Refactor imports to use direct paths for client and server modules; update API route definitions and handlers for compatibility with client-side frameworks.
